パスやクエリ文字列を指定してDBの情報を取得するAPIを実装
